+++ Assets/Scripts/App/Model/IReadOnlyCard.cs
@@ -5,7 +5,7 @@ namespace Loom.ZombieBattleground.Data
 {
     public interface IReadOnlyCard
     {
-        long MouldId { get; }
+        int MouldId { get; }
 
         string Name { get; }
