+++ Assets/Scripts/App/UI/CheatUI.cs
@@ -229,7 +229,7 @@ namespace Loom.ZombieBattleground
                                         {
                                             if (!customDeck.Cards.Any(deckCard => deckCard.MouldId == card.MouldId))
                                             {
-                                                DeckCardData deckCardData = new DeckCardData((int)card.MouldId, 0);
+                                                DeckCardData deckCardData = new DeckCardData(card.MouldId, 0);
                                                 customDeck.Cards.Add(deckCardData);
                                             }
                                         }
